LUCENE-7788: fail precommit on unparameterised log messages and examine for wasted work/objects

+++ solr/core/src/test/org/apache/solr/cloud/TestStressInPlaceUpdates.java
@@ -133,11 +133,13 @@ public class TestStressInPlaceUpdates extends AbstractFullDistribZkTestBase {
      int fullUpdatePercent = 20;
      **/
 
-    log.info("{}", Arrays.asList
-             ("commitPercent", commitPercent, "softCommitPercent", softCommitPercent,
+    if (log.isInfoEnabled()) {
+      log.info("{}", Arrays.asList
+          ("commitPercent", commitPercent, "softCommitPercent", softCommitPercent,
               "deletePercent", deletePercent, "deleteByQueryPercent", deleteByQueryPercent,
               "ndocs", ndocs, "nWriteThreads", nWriteThreads, "percentRealtimeQuery", percentRealtimeQuery,
               "operations", operations, "nReadThreads", nReadThreads));
+    }
 
     initModel(ndocs);
 
@@ -217,8 +219,8 @@ public class TestStressInPlaceUpdates extends AbstractFullDistribZkTestBase {
 
                 try {
                   returnedVersion = deleteDocAndGetVersion(Integer.toString(id), params("_version_", Long.toString(info.version)), dbq);
-                  log.info(delType + ": Deleting id=" + id + ", version=" + info.version 
-                           + ".  Returned version=" + returnedVersion);
+                  log.info("{}: Deleting id={}, version={}. Returned version={}"
+                      , delType, id, info.version, returnedVersion);
                 } catch (RuntimeException e) {
                   if (e.getMessage() != null && e.getMessage().contains("version conflict")
                       || e.getMessage() != null && e.getMessage().contains("Conflict")) {
@@ -253,7 +255,8 @@ public class TestStressInPlaceUpdates extends AbstractFullDistribZkTestBase {
                   nextVal2 = nextVal1 * 1000000000l;
                   try {
                     returnedVersion = addDocAndGetVersion("id", id, "title_s", "title" + id, "val1_i_dvo", nextVal1, "val2_l_dvo", nextVal2, "_version_", info.version);
-                    log.info("FULL: Writing id=" + id + ", val=[" + nextVal1 + "," + nextVal2 + "], version=" + info.version + ", Prev was=[" + val1 + "," + val2 + "].  Returned version=" + returnedVersion);
+                    log.info("FULL: Writing id={}, val=[{},{}], version={}, Prev was=[{},{}].  Returned version={}"
+                        ,id, nextVal1, nextVal2, info.version, val1, val2, returnedVersion);
 
                   } catch (RuntimeException e) {
                     if (e.getMessage() != null && e.getMessage().contains("version conflict")
@@ -270,7 +273,8 @@ public class TestStressInPlaceUpdates extends AbstractFullDistribZkTestBase {
                   nextVal2 = val2 + val1;
                   try {
                     returnedVersion = addDocAndGetVersion("id", id, "val2_l_dvo", map("inc", String.valueOf(val1)), "_version_", info.version);
-                    log.info("PARTIAL: Writing id=" + id + ", val=[" + nextVal1 + "," + nextVal2 + "], version=" + info.version + ", Prev was=[" + val1 + "," + val2 + "].  Returned version=" + returnedVersion);
+                    log.info("PARTIAL: Writing id={}, val=[{},{}], version={}, Prev was=[{},{}].  Returned version={}"
+                        ,id, nextVal1, nextVal2, info.version, val1, val2,  returnedVersion);
                   } catch (RuntimeException e) {
                     if (e.getMessage() != null && e.getMessage().contains("version conflict")
                         || e.getMessage() != null && e.getMessage().contains("Conflict")) {
